requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-cors-headers==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.2
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.3
Werkzeug==0.15.2
whitenoise==4.1.3
wrapt==1.11.2
psycopg2-binary==2.8.3